requirements.txt
fastapi
uvicorn
paho-mqtt
pyyaml
numpy
opencv-python-headless
faiss-cpu
requests
aiofiles